py/開発/requirements.txt
google-cloud-bigquery==1.24.0
google-cloud-logging
